backend/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["mantrakp <mantrakp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.2"
python-dotenv = "^1.0.1"
uvicorn = "^0.31.1"
fastapi-users = {extras = ["beanie", "oauth", "redis"], version = "^13.0.0"}
pymongo = "<4.9"
jinja2 = "^3.1.4"
fastapi-mail = "^1.4.1"
python-multipart = "0.0.9"
gradio = "3.50.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
